move dumps/loads to serialize.pickle
