+++ distributed/tests/test_worker.py
@@ -18,7 +18,7 @@ from distributed.batched import BatchedStream
 from distributed.core import rpc, connect, read, write
 from distributed.client import _wait
 from distributed.scheduler import Scheduler
-from distributed.serialize import dumps, loads
+from distributed.serialize.pickle import dumps, loads
 from distributed.sizeof import sizeof
 from distributed.worker import Worker, error_message, logger
 from distributed.utils import ignoring
